+++ spec/cli/fixtures/custom/a.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+# NOTE: A comment
+2 => a
+
+b := 1
+
+puts "Sum: #{a + b}"